Finish commit

+++ todolist_app/views.py
@@ -3,22 +3,26 @@ from django.core.paginator import Paginator
 from django.shortcuts import render, redirect
 from todolist_app.models import TaskList
 from todolist_app.forms import TaskForm
+from django.contrib.auth.decorators import login_required
 
 
+@login_required
 def todolist(request):
     if request.method == "POST":
         form = TaskForm(request.POST or None)
         if form.is_valid():
+            form.save(commit=False).manager = request.user
             form.save()
         messages.success(request, "New Task Added!")
         return redirect('todolist')
     else:
-        all_tasks = TaskList.objects.all()
+        all_tasks = TaskList.objects.filter(manager=request.user)
         paginator = Paginator(all_tasks, 6)
         page = request.GET.get('pg')
         all_tasks = paginator.get_page(page)
 
         return render(request, 'todolist.html', {'all_tasks': all_tasks})
+
 
 def index(request):
     context = {
@@ -34,13 +38,19 @@ def contact(request):
 def about(request):
     return render(request, 'about.html', {'about_info': 'About This App'})
 
+
+@login_required
 def delete_task(request, task_id):
     task = TaskList.objects.get(pk=task_id)
-    task.delete()
-    messages.success(request, "Task Deleted!")
+    if task.manager == request.user:
+        task.delete()
+    else:
+        messages.error(request, "Access Restricted, You Are Not Allowed!")
 
     return redirect('todolist')
 
+
+@login_required
 def edit_task(request, task_id):
     if request.method == "POST":
         task = TaskList.objects.get(pk=task_id)
@@ -53,13 +63,20 @@ def edit_task(request, task_id):
         task_obj = TaskList.objects.get(pk=task_id)
         return render(request, 'edit.html', {'task_obj': task_obj})
 
+
+@login_required
 def complete_task(request, task_id):
     task = TaskList.objects.get(pk=task_id)
-    task.done = True
-    task.save()
+    if task.manager == request.user:
+        task.done = True
+        task.save()
+    else:
+        messages.error(request, "Access Restricted, You Are Not Allowed!")
 
     return redirect('todolist')
 
+
+@login_required
 def pending_task(request, task_id):
     task = TaskList.objects.get(pk=task_id)
     task.done = False
